requirements.txt
django==3.1.2
psycopg2==2.8.6
django-tastypie==0.14.3